Fix round reputation to zero bug
Because of the generalization of our gating conditions also the
reputation conversion logic got changed. In this case we converted an
f64 into a u64 which was zero for values smaller than one. We now do the
calculations correctly by scaling up the f64 with a factor that is very
large but small enough to not overflow with a 100 percent reputation
threshold, which is the maximum the user can enter.

+++ backend/src/cli.rs
@@ -227,7 +227,7 @@ pub enum GateCmd {
         domain_id: u64,
         /// The percentage of reputation needed to get the role
         #[clap(value_hint = ValueHint::Other)]
-        reputation: u32,
+        reputation: f64,
         /// The discord role id
         #[clap(value_hint = ValueHint::Other)]
         role_id: u64,
@@ -245,7 +245,7 @@ pub enum GateCmd {
         domain_id: u64,
         /// The percentage of reputation needed to get the role
         #[clap(value_hint = ValueHint::Other)]
-        reputation: u32,
+        reputation: f64,
         /// The discord role id
         #[clap(value_hint = ValueHint::Other)]
         role_id: u64,
